The Middle
Well, any progress is good progress I guess. The systems to save work, loading however will cause the whole game to freeze forever. So thats fun. I've also looked into the systems to keep the game running while its paused, but it might be difficult thanks to how the systems with characters are set up. I'll try and polish this off later and get it all done, for now, have an unfinished and broken system.